Organizing email validation code
