--- src/database/models/email.js
@@ -1,8 +1,0 @@
-const mongoose = require('mongoose');
-
-const email = mongoose.Schema({
-  email: String,
-  confirmationCode: String,
-});
-
-module.exports = mongoose.model('Email', email);
